hwclock: size optimizations libbb/time.c: new file, introducing monotonic_us() pscan, traceroute, arping: use it instead of gettimeofday ping, zcip: TODO
function                                             old     new   delta
monotonic_us                                           -      89     +89
find_pair                                            164     180     +16
.rodata                                           129747  129763     +16
refresh                                             1144    1152      +8
............
timeout                                                8       4      -4
static.start                                           8       4      -4
last                                                   8       4      -4
parse_conf                                          1303    1284     -19
time_main                                           1149    1124     -25
gettimeofday_us                                       39       -     -39
arping_main                                         2042    1969     -73
hwclock_main                                         594     501     -93
catcher                                              485     380    -105
traceroute_main                                     4300    4117    -183
------------------------------------------------------------------------------
(add/remove: 2/1 grow/shrink: 8/11 up/down: 157/-562)        Total: -405 bytes

+++ networking/wget.c
@@ -705,7 +705,7 @@ progressmeter(int flag)
 	ratio = 100;
 	if (totalsize != 0 && !chunked) {
 		/* long long helps to have working ETA even if !LFS */
-		ratio = (int) (100 * (unsigned long long)(transferred+beg_range) / totalsize);
+		ratio = (int) (100ULL * (transferred+beg_range) / totalsize);
 		ratio = MIN(ratio, 100);
 	}
